requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv
pandas==2.1.4
selenium==4.15.2
webdriver-manager==4.0.1
google-generativeai==0.3.2
python-multipart==0.0.6
aiofiles==23.2.1
requests==2.31.0
